+++ source/extensions/lightspeed.trex.stage_manager.plugin.filter.usd/lightspeed/trex/stage_manager/plugin/filter/usd/is_logic_graph.py
@@ -15,20 +15,21 @@
 * limitations under the License.
 """
 
-__all__ = ["RemixLogicPrimsFilterPlugin"]
+__all__ = ["RemixLogicPrimsFilterPlugin", "FilterTypes"]
 
-from typing import TYPE_CHECKING
+from enum import Enum
 
-from lightspeed.common.constants import OMNI_GRAPH_NODE_TYPE, OMNI_GRAPH_TYPE
+from lightspeed.common.constants import OMNI_GRAPH_NODE_TYPES, OMNI_GRAPH_TYPE
 from omni import ui
 from omni.flux.stage_manager.factory import StageManagerItem
 from omni.flux.stage_manager.plugin.filter.usd.base import StageManagerUSDFilterPlugin
 from pydantic import Field, PrivateAttr
 
-if TYPE_CHECKING:
-    from omni.flux.stage_manager.factory.plugins.tree_plugin import StageManagerTreeItem, StageManagerTreeModel
 
-_filter_types = ["No Filter", "Graphs Only", "Graphs + Nodes"]
+class FilterTypes(Enum):
+    NO_FILTERS = "No Filter"
+    GRAPHS_ONLY = "Graphs Only"
+    GRAPHS_AND_NODES = "Graphs + Nodes"
 
 
 class RemixLogicPrimsFilterPlugin(StageManagerUSDFilterPlugin):
@@ -38,17 +39,20 @@ class RemixLogicPrimsFilterPlugin(StageManagerUSDFilterPlugin):
     _COMBO_BOX_WIDTH: int = PrivateAttr(default=130)
     _filter_combobox: ui.ComboBox | None = PrivateAttr(default=None)
     _current_index: int = PrivateAttr(default=0)
-    current_filter_type: str = Field(default="No Filter", description="The type of logic to filter by")
+    current_filter_type: FilterTypes = Field(
+        default=FilterTypes.NO_FILTERS, description="The type of logic to filter by"
+    )
 
     def filter_predicate(self, item: StageManagerItem) -> bool:
-        match self._current_index:
-            case 1:
+        match self.current_filter_type:
+            case FilterTypes.GRAPHS_ONLY:
                 return item.data.GetTypeName() == OMNI_GRAPH_TYPE
-            case 2:
-                return item.data.GetTypeName() in {OMNI_GRAPH_TYPE, OMNI_GRAPH_NODE_TYPE}
+            case FilterTypes.GRAPHS_AND_NODES:
+                return item.data.GetTypeName() in OMNI_GRAPH_NODE_TYPES
         return True
 
     def build_ui(self) -> None:
+        filter_labels = [each.value for each in FilterTypes]
         with ui.HStack(spacing=ui.Pixel(8)):
             ui.Spacer(width=0)
             ui.Label(
@@ -58,13 +62,18 @@ class RemixLogicPrimsFilterPlugin(StageManagerUSDFilterPlugin):
             )
             self._filter_combobox = ui.ComboBox(
                 self._current_index,
-                *_filter_types,
+                *filter_labels,
                 width=ui.Pixel(self._COMBO_BOX_WIDTH),
             )
             self._filter_combobox.model.add_item_changed_fn(self._on_filter_changed)
 
-    def _on_filter_changed(self, model: "StageManagerTreeModel", _: "StageManagerTreeItem") -> None:
+    def _on_filter_changed(self, model: ui.AbstractItemModel, _: ui.AbstractItem) -> None:
         selected_index = model.get_item_value_model().get_value_as_int()
         self._current_index = selected_index
-        self.current_filter_type = _filter_types[selected_index]
+        self.current_filter_type = list(FilterTypes)[selected_index]
         self._filter_items_changed()
+
+    def model_post_init(self, *args, **kwargs) -> None:
+        super().model_post_init(*args, **kwargs)
+        if isinstance(self.current_filter_type, FilterTypes):
+            self._current_index = list(FilterTypes).index(self.current_filter_type)