+++ source/extensions/lightspeed.layer_manager.core/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [3.1.0]
+### Added
+- Added `get_all_replacement_layers` to list all replacement layers in the layer hierarchy
+
 ## [3.0.3]
 ### Added
 - Added `close_project_with_data_models` method to close stage via context