+++ source/extensions/lightspeed.trex.menu.workfile/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.5.1]
+### Fixed
+- Fixed save dialog error
+
 ## [1.5.0]
 ### Changed
 - Migrated from custom burger menu to standard Kit menubar integration